fix 1 vài loi cong thuc và làm khuyen mai chua ktra dieu kien

+++ src/main/java/com/example/retaurant/DAO/CongThucDAO.java
@@ -89,7 +89,7 @@ public class CongThucDAO {
         CongThucDTO ct = new CongThucDTO();
         ct.setSpid(resultSet.getInt("sp_id"));
         ct.setNlid(resultSet.getInt("nl_id"));
-        ct.setSoluong(resultSet.getInt("soLuong"));
+        ct.setSoluong(resultSet.getFloat("soLuong"));
         return ct;
     }
